done some errors

+++ src/main/java/studio/demo/service/impl/PromotionServiceImpl.java
@@ -2,7 +2,13 @@ package studio.demo.service.impl;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import studio.demo.exception.*;
+import studio.demo.model.binding.CommentAddBindingModel;
+import studio.demo.model.binding.PromotionAddBindingModel;
+import studio.demo.model.entity.Comment;
 import studio.demo.model.entity.Promotion;
+import studio.demo.model.entity.User;
+import studio.demo.model.service.CommentServiceModel;
 import studio.demo.model.service.PromotionServiceModel;
 import studio.demo.model.view.PromotionViewModel;
 import studio.demo.repository.PromotionRepository;
@@ -49,15 +55,58 @@ public class PromotionServiceImpl implements PromotionService {
     }
 
 
+
     @Override
-    public void delete(String id) {
+    public boolean delete(String id) throws PromotionWithThisIdDoesNotExist {
+        if (this.promotionRepository.findById(id).isEmpty()) {
+            throw new PromotionWithThisIdDoesNotExist("Promotion with this id does not exist.");
+        }
         this.promotionRepository.deleteById(id);
+        return true;
     }
 
     @Override
-    public void addPromotion(PromotionServiceModel promotionServiceModel) {
+    public PromotionServiceModel addPromotion(PromotionServiceModel promotionServiceModel) throws PromotionDoesNotExist, PromotiontWithThisNameExist {
+        Promotion ppp=promotionRepository.findByName(promotionServiceModel.getName()).orElse(null);
+
+        if (promotionServiceModel.getName().trim().isEmpty()) {
+            throw new PromotionDoesNotExist("Promotion is null.");
+        }
+        if (ppp != null) {
+            throw new PromotiontWithThisNameExist("Promotion with this topic exists!");
+        }
         Promotion promotion = this.modelMapper.map(promotionServiceModel, Promotion.class);
 
         this.promotionRepository.saveAndFlush(promotion);
+
+
+        return promotionServiceModel;
     }
+
+
+    @Override
+    public PromotionServiceModel update(PromotionAddBindingModel promotion) throws PromotionWithThisNameDoesNotExist {
+
+        Promotion ppp = this.promotionRepository.
+                findByName(promotion.getName()).orElse(null);
+        this.checkPromotionExist(ppp);
+
+        if (!promotion.getDescription().trim().isEmpty()) {
+            ppp.setDescription(promotion.getDescription());
+        }
+        if (promotion.getPrice() != null) {
+            ppp.setPrice(promotion.getPrice());
+        }
+        return this.modelMapper.map(this.promotionRepository.saveAndFlush(ppp),
+                PromotionServiceModel.class);
+
+    }
+
+    private void checkPromotionExist(Promotion promotion) throws PromotionWithThisNameDoesNotExist {
+        if (promotion == null) {
+            throw new PromotionWithThisNameDoesNotExist("Promotion with this topic does not exist!");
+        }
+    }
+
+
 }
